Extract input.go
